+++ daisygame/js/values.js
@@ -22,6 +22,7 @@ const MODE_ENDLESS = 2;
 const MODE_ARCADE_KEY = 201;
 const MODE_PUZZLE_KEY = 202;
 const MODE_ENDLESS_KEY = 203;
+const MENU_KEY = 204;
 
 let gStartX = 0;
 let gBlockSize = 60;